+++ src/conduit-lwt-unix/conduit_lwt_unix.mli
@@ -160,13 +160,17 @@ val default_ctx : ctx
 val init :
   ?src:string ->
   ?tls_own_key:tls_own_key ->
-  ?tls_server_key:tls_own_key (* Deprecated, use tls_own_key. *) ->
+  ?tls_authenticator:Conduit_lwt_tls.X509.authenticator ->
   unit ->
   ctx io
 (** [init ?src ?tls_own_key ()] will initialize a Unix conduit that binds to the
     [src] interface if specified. If TLS server connections are used, then
     [tls_server_key] must contain a valid certificate to be used to advertise a
-    TLS connection *)
+    TLS connection.
+
+    The certificate is validated using [tls_authenticator]. By default, the
+    validation is using the {{:https://github.com/mirage/ca-certs} OS trust
+    anchors}. *)
 
 val connect : ctx:ctx -> client -> (flow * ic * oc) io
 (** [connect ~ctx client] establishes an outgoing connection via the [ctx]